refactored policies into policy package
made a policy package. Added support for continuous actions and joystick clipping
