+++ tests/prt_rl/utils/policy/test_random_policy.py
@@ -1,0 +1,78 @@
+import torch
+from tensordict.tensordict import TensorDict
+from prt_rl.env.interface import EnvParams
+from prt_rl.utils.policy import RandomPolicy
+
+def test_random_discrete_action_selection():
+    # Create a fake environment that has 1 discrete action [0,...,3] and 1 discrete state with the same interval
+    params = EnvParams(
+        action_shape=(1,),
+        action_continuous=False,
+        action_min=0,
+        action_max=3,
+        observation_shape=(1,),
+        observation_continuous=False,
+        observation_min=0,
+        observation_max=3,
+    )
+
+    policy = RandomPolicy(env_params=params)
+
+    # Set seed for consistent unit tests
+    torch.manual_seed(3)
+
+    # Create fake observation TensorDict
+    td = TensorDict({}, batch_size=[1])
+    td = policy.get_action(td)
+
+    assert td['action'].shape == (1, 1)
+    assert td['action'][0] == 2
+
+
+def test_random_continuous_action_selection():
+    params = EnvParams(
+        action_shape=(1,),
+        action_continuous=True,
+        action_min=[1.0],
+        action_max=[1.1],
+        observation_shape=(1,),
+        observation_continuous=False,
+        observation_min=0,
+        observation_max=3,
+    )
+
+    policy = RandomPolicy(env_params=params)
+
+    # Set seed for consistent unit tests
+    torch.manual_seed(0)
+
+    # Create fake observation TensorDict
+    td = TensorDict({}, batch_size=[1])
+    td = policy.get_action(td)
+
+    assert td['action'].shape == (1, 1)
+    assert torch.allclose(td['action'], torch.tensor([[1.05]]), atol=1e-2)
+
+def test_random_multiple_continuous_action_selection():
+    params = EnvParams(
+        action_shape=(3,),
+        action_continuous=True,
+        action_min=[0.0, 0.0, 0.0],
+        action_max=[1.0, 1.0, 1.0],
+        observation_shape=(1,),
+        observation_continuous=False,
+        observation_min=0,
+        observation_max=3,
+    )
+
+    policy = RandomPolicy(env_params=params)
+
+    # Set seed for consistent unit tests
+    torch.manual_seed(0)
+
+    # Create fake observation TensorDict
+    td = TensorDict({}, batch_size=[1])
+    td = policy.get_action(td)
+
+    assert td['action'].shape == (1, 3)
+    assert torch.allclose(td['action'], torch.tensor([[0.50, 0.77, 0.09]]), atol=1e-2)
